user routing merged
