+++ forum/routes.py
@@ -13,56 +13,56 @@ from .user import username_taken, email_taken, valid_username, User
 
 rt = Blueprint('routes', __name__, template_folder='templates')
 
-@rt.route('/action_login', methods=['POST'])
-def action_login():
-	# Read login form values and authenticate the user.
-	username = request.form['username']
-	password = request.form['password']
-	user = User.query.filter(User.username == username).first()
-	if user and user.check_password(password):
-		login_user(user)
-	else:
-		errors = []
-		errors.append("Username or password is incorrect!")
-		return render_template("login.html", errors=errors)
-	return redirect("/")
+# @rt.route('/action_login', methods=['POST'])
+# def action_login():
+# 	# Read login form values and authenticate the user.
+# 	username = request.form['username']
+# 	password = request.form['password']
+# 	user = User.query.filter(User.username == username).first()
+# 	if user and user.check_password(password):
+# 		login_user(user)
+# 	else:
+# 		errors = []
+# 		errors.append("Username or password is incorrect!")
+# 		return render_template("login.html", errors=errors)
+# 	return redirect("/")
 
-@login_required
-@rt.route('/action_logout')
-def action_logout():
-	# End the current session and send the user back home.
-	logout_user()
-	return redirect("/")
+# @login_required
+# @rt.route('/action_logout')
+# def action_logout():
+# 	# End the current session and send the user back home.
+# 	logout_user()
+# 	return redirect("/")
 
-@rt.route('/action_createaccount', methods=['POST'])
-def action_createaccount():
-	# Validate signup data, create the user, then log them in.
-	username = request.form['username']
-	password = request.form['password']
-	email = request.form['email']
-	errors = []
-	retry = False
-	if username_taken(username):
-		errors.append("Username is already taken!")
-		retry = True
-	if email_taken(email):
-		errors.append("An account already exists with this email!")
-		retry = True
-	if not valid_username(username):
-		errors.append("Username is not valid!")
-		retry = True
-	# if not valid_password(password):
-	# 	errors.append("Password is not valid!")
-	# 	retry = True
-	if retry:
-		return render_template("login.html", errors=errors)
-	user = User(email, username, password)
-	if user.username == "admin":
-		user.admin = True
-	db.session.add(user)
-	db.session.commit()
-	login_user(user)
-	return redirect("/")
+# @rt.route('/action_createaccount', methods=['POST'])
+# def action_createaccount():
+# 	# Validate signup data, create the user, then log them in.
+# 	username = request.form['username']
+# 	password = request.form['password']
+# 	email = request.form['email']
+# 	errors = []
+# 	retry = False
+# 	if username_taken(username):
+# 		errors.append("Username is already taken!")
+# 		retry = True
+# 	if email_taken(email):
+# 		errors.append("An account already exists with this email!")
+# 		retry = True
+# 	if not valid_username(username):
+# 		errors.append("Username is not valid!")
+# 		retry = True
+# 	# if not valid_password(password):
+# 	# 	errors.append("Password is not valid!")
+# 	# 	retry = True
+# 	if retry:
+# 		return render_template("login.html", errors=errors)
+# 	user = User(email, username, password)
+# 	if user.username == "admin":
+# 		user.admin = True
+# 	db.session.add(user)
+# 	db.session.commit()
+# 	login_user(user)
+# 	return redirect("/")
 
 
 # @rt.route('/subforum')
@@ -81,10 +81,10 @@ def action_createaccount():
 # 	subforums = Subforum.query.filter(Subforum.parent_id == subforum_id).all()
 # 	return render_template("subforum.html", subforum=subforum, posts=posts, subforums=subforums, path=subforumpath)
 
-@rt.route('/loginform')
-def loginform():
-	# Render the shared login and signup page.
-	return render_template("login.html")
+# @rt.route('/loginform')
+# def loginform():
+# 	# Render the shared login and signup page.
+# 	return render_template("login.html")
 
 
 # @login_required
